pyproject.toml
[tool.poetry]
name = "ducklens"
version = "0.1.0"
description = "AI-powered stock price prediction system"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = ">=0.104.0"
uvicorn = {extras = ["standard"], version = ">=0.24.0"}
pydantic = ">=2.5.0"
pydantic-settings = ">=2.1.0"
duckdb = ">=0.9.0"
supabase = ">=2.0.0"
pandas = ">=2.1.0"
numpy = ">=1.26.0"
scikit-learn = ">=1.3.0"
httpx = ">=0.25.0"
polygon-api-client = ">=1.12.0"
python-dotenv = ">=1.0.0"
loguru = ">=0.7.0"
tenacity = ">=8.2.0"
colorama = ">=0.4.6"
rich = ">=13.0.0"
plotille = ">=5.0.0"
asciichartpy = ">=1.5.25"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.0"
pytest-asyncio = ">=0.21.0"
pytest-cov = ">=4.1.0"
mypy = ">=1.7.0"
ruff = ">=0.1.0"
black = ">=23.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
select = ["E", "W", "F", "I", "B", "C4", "UP", "ANN"]
ignore = ["E501", "ANN101"]
line-length = 100
target-version = "py311"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = ["--strict-markers", "--cov=src", "--cov-report=term-missing", "-v"]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
